Send requests for driving anf rail to maps API
Note that currently the data from the maps API is not being processed
and only the tests for `DistanceController` are guaranteed to pass in
this commit.

We have refactored api calls into separate methods to make the code
more declarative, and with a view of using Promise.all to run all
requests concurrently.